scripts/scenario-runner/pyproject.toml
[build-system]
requires = ["setuptools>=64"]
build-backend = "setuptools.build_meta"

[project]
name = "scenario-runner"
version = "0.1.0"
description = "YAML-driven scenario testing for vykar"
requires-python = ">=3.9"
dependencies = [
    "faker>=18.0",
    "faker-file>=0.17",
    "PyYAML>=6.0",
    "python-docx>=1.0",
    "tablib[xlsx]>=3.5",
]

[tool.setuptools]
packages = ["scenario_runner"]

[project.scripts]
scenario-runner = "scenario_runner.cli:main"
